iOS Scalebar: Move Windows and Xamarin-specific logic into platform-specific partial class files
