+++ src/Esri.ArcGISRuntime.Toolkit/UWP/Esri.ArcGISRuntime.Toolkit.UWP.csproj
@@ -91,6 +91,7 @@
     </Page>
   </ItemGroup>
   <Import Project="..\Shared\All\Esri.ArcGISRuntime.Toolkit.Shared.projitems" Label="Shared" />
+  <Import Project="..\Shared\Windows\Esri.ArcGISRuntime.Toolkit.Windows.Shared.projitems" Label="Shared" />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
   </PropertyGroup>
